app/requirements.txt
gradio==4.13.0
google-cloud-aiplatform==1.38.1
python-dotenv==1.0.0
pillow==10.2.0
python-dotenv==1.0.0